RatKing Animations in and plugged into their attacks
The animations are in and plugged into their respective attacks. It looks like the IK is off on them so the feet aren't currently detecting the floor properly at least for the animation purposes but I've adjusting them so that the RatKing is no longer floating during the attack animations.