Sunei & Juice
Sunei animations are in and connected to their respective trigger boxes, rats and the Ratking have a material swap to show when they're damaged.
Particle system for one of the Sunei appearances and a Level particle system that is currently only in the bedroom. And some light file organization via folders